requirements-dev.txt
babel==2.3.4
mock==2.0.0
nose2[coverage-plugin]
pep8==1.7.0
transifex-client==0.12.2
unittest2==1.1.0
zope.testrunner==4.5.1
pytest-splinter==1.7.8
